Support for ESLint v8

+++ tests/dep-time-travel.sh
@@ -8,19 +8,23 @@ export NPM_CONFIG_LEGACY_PEER_DEPS=true
 
 npm install --no-save "eslint@${ESLINT_VERSION}" --ignore-scripts
 
-# completely remove the new TypeScript parser for ESLint < v5
-if [[ "$ESLINT_VERSION" -lt "5" ]]; then
-  echo "Removing @typescript-eslint/parser..."
-  npm uninstall --no-save @typescript-eslint/parser
-fi
+re="^[0-9]+$"
 
-# use these alternate TypeScript dependencies for ESLint < v4
-if [[ "$ESLINT_VERSION" -lt "4" ]]; then
-  echo "Downgrading babel-eslint..."
-  npm i --no-save babel-eslint@8.0.3
+if [[ "$ESLINT_VERSION" =~ $re ]] ; then
+  # completely remove the new TypeScript parser for ESLint < v5
+  if [[ "$ESLINT_VERSION" -lt "5" ]]; then
+    echo "Removing @typescript-eslint/parser..."
+    npm uninstall --no-save @typescript-eslint/parser
+  fi
 
-  echo "Downgrading TypeScript dependencies..."
-  npm i --no-save typescript-eslint-parser@15 typescript@2.8.1
+  # use these alternate TypeScript dependencies for ESLint < v4
+  if [[ "$ESLINT_VERSION" -lt "4" ]]; then
+    echo "Downgrading babel-eslint..."
+    npm i --no-save babel-eslint@8.0.3
+
+    echo "Downgrading TypeScript dependencies..."
+    npm i --no-save typescript-eslint-parser@15 typescript@2.8.1
+  fi
 fi
 
 # typescript-eslint-parser 1.1.1+ is not compatible with node 6
